Issue 2255 - Loading very large MPAS files results in an abort  (#2256)
* Removed duplicated Transpose code

* Made Wasp::Transpose() functions 64bit clean

* Added Initialize() method to RenderParams. Previously, there was no
initialization method and the RenderParams constructor was not no-fail.
In particular, an assert was thrown if the constructor could not determine
the extents of a variable from the DataMgr. Getting variable extents is
an operation that can fail for a number of reasons such as running out of
memory.

* Fix needed for renderers that support both 2D and 3D variables

* Fixed regression - Added Initialize() method to SliceParams, ImageParams,
and FlowParams.

* Update include/vapor/FlowParams.h


* Update include/vapor/ImageParams.h


* Update include/vapor/SliceParams.h


* Fixed regression:

Even though vapor fails to open the dataset, vapor enters a limbo
state where a dataset is both open and not open. The GUI indicates
no dataset is open however if you try opening another dataset, you
will get a prompt asking whether you want to Load as new dataset
or replace existing?. I am assuming this is a bug.

* Fixing warnings generated by adding "override" qualifiers to
RenderParams::Initialize() method. The fix was to simply remove the
unused RenderParams IsOpaque() and usingVariable() methods.

* Moved GetAppRenderParams implementation to .cpp file

+++ include/vapor/utils.h
@@ -140,7 +140,8 @@ COMMON_API std::vector<size_t> VectorizeCoords(
 //    s1,s2: size of entire matrix (row,col)
 //
 COMMON_API void Transpose(
-    const float *a, float *b, int p1, int m1, int s1, int p2, int m2, int s2);
+    const float *a, float *b, size_t p1, size_t m1, size_t s1,
+    size_t p2, size_t m2, size_t s2);
 
 //
 // blocked matrix Transpose single threaded
@@ -148,7 +149,7 @@ COMMON_API void Transpose(
 //   *b : pointer to output matrix
 //    s1,s2: size of entire matrix (row,col)
 //
-COMMON_API void Transpose(const float *a, float *b, int s1, int s2);
+COMMON_API void Transpose(const float *a, float *b, size_t s1, size_t s2);
 
 // Perform a binary search in a sorted (increasing or decreasing) 1D
 // vector of values for the
